ssize_t => int64_t

+++ include/slurp.h
@@ -221,7 +221,7 @@ int slurp_xz(slurp_t *src);
 int slurp_zstd(slurp_t *src);
 #endif
 
-int slurp_is_valid_file_pointer(slurp_t *fp, ssize_t x, int whence);
+int slurp_is_valid_file_pointer(slurp_t *fp, int64_t x, int whence);
 
 void slurp_fill_nonseek_buffer(slurp_t *t, size_t count);
